+++ @nodepy/pm/lib/manifest.py
@@ -70,23 +70,10 @@ class PackageVersion:
 
 class PackageManifest:
   """
-  This class describes a `package.json` package manifest in memory. The
-  manifest can have the following fields:
-
-  - name (required)
-  - version (required)
-  - license
-  - description
-  - repository
-  - bin
-  - script
-  - engines
-  - dependencies
-  - python-dependencies
-  - dist
-
-  Additional fields are only accepted when the name of that field also appears
-  in the *engines* object, for example
+  This class describes a `package.json` package manifest in memory. Check out
+  the #schema for a description of supported fields. Additional fields are
+  only accepted when the name of that field also appears in the *engines*
+  object, for example
 
   ```json
   {
@@ -115,6 +102,7 @@ class PackageManifest:
       "repository": {"type": "string"},
       "license": {"type": "string"},
       "private": {"type": "boolean"},
+      "main": {"type": "string"},
       "dependencies": {
         "type": "object",
         "additionalProperties": {"type": "string"}
@@ -162,7 +150,7 @@ class PackageManifest:
       author=None, license=None, dependencies=None, dev_dependencies=None,
       python_dependencies=None, dev_python_dependencies=None, scripts=None,
       bin=None, engines=None, engine_props=None, dist=None, repository=None,
-      private=False):
+      private=False, main=None):
     if len(name) < 2 or len(name) > 127:
       raise ValueError('packag name must be at least 2 and maximum 127 characters')
     if name.startswith('_') or name.startswith('.'):
@@ -189,6 +177,7 @@ class PackageManifest:
     self.engine_props = {} if engine_props is None else engine_props
     self.dist = {} if dist is None else dist
     self.private = private
+    self.main = main
 
   def __eq__(self, other):
     if isinstance(other, PackageManifest):